requirements.txt
psycopg[binary]>=3.1.0
watchdog>=4.0.0
python-dotenv>=1.0.0
